+++ wannago-server/src/main/java/com/wannago/qna/answer/service/AnswerServiceImpl.java
@@ -7,8 +7,8 @@ import com.wannago.member.repository.MemberRepository;
 import com.wannago.common.exception.CustomErrorCode;
 import com.wannago.common.exception.CustomException;
 import com.wannago.qna.answer.dto.AnswerResponse;
-import com.wannago.qna.ask.entity.Answer;
-import com.wannago.qna.ask.entity.Ask;
+import com.wannago.qna.entity.Answer;
+import com.wannago.qna.entity.Ask;
 import com.wannago.qna.ask.repository.AskRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
