+++ wannago-server/src/main/java/com/wannago/qna/entity/Category.java
@@ -1,4 +1,4 @@
-package com.wannago.qna.ask.entity;
+package com.wannago.qna.entity;
 
 public enum Category {
     ACTIVITY, FOOD, LOCATION, OTHER;
